refactor(web): co-locate PageIndex credential form
