+++ web/hooks/useKnowledgeBases.ts
@@ -176,7 +176,6 @@ export function useKnowledgeBases() {
       name: string;
       provider: string;
       files: File[];
-      pageindexMode?: "flash" | "standard";
     }): Promise<KnowledgeTaskResponse> => {
       const result = await createKbApi(params);
       invalidateKnowledgeCaches();
